// @starbeam/core/strippable/trace.d.ts
export declare class Group {
    #private;
    static start(console: TraceConsole, label: string, shouldLog: boolean): Group;
    private constructor();
    collapsed(): Group;
    collapsed<T>(callback: () => T): T;
    expanded(): Group;
    end(): void;
}
interface TraceModifiers {
    readonly withStack: TraceMethods;
}
interface TraceMethods extends TraceModifiers {
    log(format: string, ...args: unknown[]): void;
    log(...args: unknown[]): void;
    log(format: string, ...args: unknown[]): void;
    log(...args: unknown[]): void;
    group(description: string): Group;
    group<T>(description: string, callback: () => T): T;
}
interface TraceLevels {
    readonly trace: TraceMethods;
    readonly warn: TraceMethods;
}
// ...
export declare const LOGGER: TraceMethods & TraceLevels;